api/pyproject.toml
[project]
name = "ib-ox-api"
version = "0.1.0"
description = "Read-only API for IB-Oxford longitudinal questionnaire data"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "click>=8.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
test = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
ib-ox-api = "ib_ox_api.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
